Removes duplicated code that was dead, marks other areas where duplication is
