+++ x-pack/plugins/security_solution/public/common/components/autocomplete/types.ts
@@ -7,20 +7,8 @@
 
 import { EuiComboBoxOptionOption } from '@elastic/eui';
 
-import type {
-  ListOperatorEnum as OperatorEnum,
-  ListOperatorTypeEnum as OperatorTypeEnum,
-} from '@kbn/securitysolution-io-ts-list-types';
-
 export interface GetGenericComboBoxPropsReturn {
   comboOptions: EuiComboBoxOptionOption[];
   labels: string[];
   selectedComboOptions: EuiComboBoxOptionOption[];
-}
-
-export interface OperatorOption {
-  message: string;
-  value: string;
-  operator: OperatorEnum;
-  type: OperatorTypeEnum;
 }
